+++ src/main/java/seedu/address/logic/parser/ArgumentMultimap.java
@@ -9,6 +9,7 @@ import java.util.stream.Stream;
 
 import seedu.address.logic.Messages;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.logic.parser.prefix.Prefix;
 
 /**
  * Stores mapping of prefixes to their respective arguments.
