+++ src/main/java/seedu/address/model/person/predicates/FilterListPredicate.java
@@ -1,9 +1,9 @@
 package seedu.address.model.person.predicates;
 
-import seedu.address.model.person.Person;
-
 import java.util.List;
 import java.util.function.Predicate;
+
+import seedu.address.model.person.Person;
 
 /**
  * A predicate that tests if a given person is present in a list of persons.
